server/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & LangChain
langchain==0.0.350
langchain-openai==0.0.2
openai==1.6.1

# Document Processing
pytesseract==0.3.10
pdfplumber==0.10.2
pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.1.3

# WebSocket
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
